+++ instrumentation/opentelemetry-instrumentation-pika/src/opentelemetry/instrumentation/pika/utils.py
@@ -27,11 +27,14 @@ class _PikaGetter(Getter):  # type: ignore
 
 _pika_getter = _PikaGetter()
 
+HookT = Optional[Callable[[Span, bytes, BasicProperties,], None,]]
+
 
 def _decorate_callback(
     callback: Callable[[Channel, Basic.Deliver, BasicProperties, bytes], Any],
     tracer: Tracer,
     task_name: str,
+    consume_hook: HookT = None,
 ):
     def decorated_callback(
         channel: Channel,
@@ -54,6 +57,8 @@ def _decorate_callback(
             operation=MessagingOperationValues.RECEIVE,
         )
         with trace.use_span(span, end_on_exit=True):
+            if callable(consume_hook):
+                consume_hook(span, body, properties)
             retval = callback(channel, method, properties, body)
         return retval
 
@@ -64,6 +69,7 @@ def _decorate_basic_publish(
     original_function: Callable[[str, str, bytes, BasicProperties, bool], Any],
     channel: Channel,
     tracer: Tracer,
+    publish_hook: HookT = None,
 ):
     def decorated_function(
         exchange: str,
@@ -91,6 +97,8 @@ def _decorate_basic_publish(
         with trace.use_span(span, end_on_exit=True):
             if span.is_recording():
                 propagate.inject(properties.headers)
+                if callable(publish_hook):
+                    publish_hook(span, body, properties)
             retval = original_function(
                 exchange, routing_key, body, properties, mandatory
             )